+++ DreamKeyAPI/src/main/java/org/generation/DreamKeyAPI/repository/AfiliadosRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AfiliadosRepository extends JpaRepository<Afiliados, Long>{
+public interface AfiliadosRepository extends JpaRepository<Afiliados, Integer>{
 	Optional<Afiliados> findByNombre(String nombre);
 }//class AfiliadosRepository
