+++ DreamKeyAPI/src/main/java/org/generation/DreamKeyAPI/repository/SuscripcionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SuscripcionRepository extends JpaRepository <Suscripcion, Long> {
+public interface SuscripcionRepository extends JpaRepository <Suscripcion, Integer> {
 
 }
